+++ CHANGELOG.rst
@@ -2,8 +2,9 @@ Release notes
 -------------
 ### Version 2.0.0
 
-2021-05-27
+2021-05-28
  - Pass a list of dictionaries when doing rendering instead of using field objects
+ - Output number of component generated in the attribution if default_html.template is used.
  
 
 ### Version 1.0.0